requirements.txt
google-api-python-client==1.7.8
requests==2.21.0
PyJWT==1.7.1